aj-empathix/candidates/requirements.txt
requests
boto3
Flask
flask-lambda-python36
Werkzeug
aws-wsgi
serverless-wsgi